BUGFIX Calling tinyMCE.init() within entwine to avoid DOMReady execution order problems.
Changed from inline script calls to tinyMCE.init() to a new ssTinyMceConfig global variable
which gets evaluated in LeftAndMain.EditForm.js. Also loads TinyMCE regardless of an actual
HTMLEditorField being present in the CMS UI, as we need to prepare for loading other UIs with TinyMCE via ajax.
Usage of HTMLEditorField outside of the CMS is not impacted, as long as the existing HTMLEditorField::include_js() call is used.